Share the beans cached context with the endpoint test
Fold AxelixBeansEndpointTest into the AbstractBeansSharedContextTest
cached context so the whole beans test package uses one Spring
ApplicationContext instead of two.

The endpoint configuration (CurrentConfiguration, AxelixPropTest, the
RANDOM_PORT web environment, the actuator endpoint imports and the test
property source) is moved into the parent so it keeps owning the shared
configuration directly and never references its subclasses.
AxelixBeansEndpointTest becomes a thin subclass with only its test
methods.

The duplicate conditionalBeanRefBuilder, beanMetaInfoExtractor and
qualifiersPersistencePostProcessor beans are removed from
DefaultBeanAnalyzerTestConfig, since they would otherwise collide with
CurrentConfiguration's beans now that both live in the same context
(bean overriding is disabled).
